Set user's CN to their displayName
In 7f7978a6be69a0663eba4f92e9c674a1fbb230fd the cn was set to be the uid
of the user. This was to be consistant with AD systems where this is the
case.

According to http://www.zytrax.com/books/ldap/ch3/, cn is the "globally
unique name for this attribute", so setting this to the uid is
technically correct because only the UID is guaranteed to be globally
unique.

Many systems however assume the cn contains the full name of the user
(first name plus last name) and as such break when cn contains the
username. So this change sets cn to contain the full name of the user
again.

LDAP also defines the display name attribute which stores the full name
of the user.